remade lectureItem to component, added personItem component
