+++ client/src/app.config.router.js
@@ -1,6 +1,6 @@
 const lecturesListController = require("./states/student/lecturesList/controller.js");
 const homeController = require("./states/student/home/controller.js");
-const personsListComponent = require('./components/personsList/personsListComponent.js');
+const personsListComponent = require('./states/student/studentList/component.js');
 
 module.exports = function(app) {
   // We can register controller with angular first
@@ -12,18 +12,10 @@ module.exports = function(app) {
   app.config(['$stateProvider', '$urlRouterProvider', function($stateProvider, $urlRouterProvider) {
     $stateProvider
 
-
+    /* STUDENT STATE */
     .state('student', {
         url: '/student',
         template: require('./states/student/index.html')
-    })
-    .state('teacher', {
-        url: '/teacher',
-        template: 'hello from personal teacher page'
-    })
-    .state('admin', {
-        url: '/admin',
-        template: 'hello from personal admin page'
     })
     // students state using component method
     // now the controller and template included in component
@@ -33,19 +25,30 @@ module.exports = function(app) {
       component: 'personsListComponent'
     })
 
-    .state('lectures', {
+    .state('student.lectures', {
       url: '/lectures',
       template: require('./states/student/lecturesList/template.html'),
       controller: lecturesListController,
       controllerAs: '$ctrl'
     })
 
-    .state('home', {
+    .state('student.home', {
       url: '/',
       template: require('./states/student/home/template.html'),
       // controller as string - name of previously registered controller
       controller: 'homeController',
       controllerAs: '$ctrl'
+    })
+
+    /* TEACHER STATE */
+    .state('teacher', {
+        url: '/teacher',
+        template: 'hello from personal teacher page'
+    })
+    /* ADMIN STATE */
+    .state('admin', {
+        url: '/admin',
+        template: 'hello from personal admin page'
     })
 
     $urlRouterProvider.otherwise('/');
