+++ client/src/components/lectureItem/lectureItem.html
@@ -1,4 +1,4 @@
-<div class="card cardLecture" ng-repeat="lecture in lectureItem.lectures">
+<div class="card cardLecture" ng-repeat="lecture in $ctrl.lectures">
   <div class="card-image waves-effect waves-block waves-light">
     <div class="row dateLine">
       <div class="col s6">
