+++ client/src/components/lectureItem/lectureItemComponent.js
@@ -1,0 +1,11 @@
+const lectureItemController = require('./lectureItemController');
+
+var lectureItemComponent = {
+    bindings: {
+        lectures: '=?'
+    },
+    template: require('./lectureItem.html'),
+    controller: lectureItemController
+};
+
+module.exports = lectureItemComponent;
