+++ client/src/components/lectureItem/lectureItemController.js
@@ -1,3 +1,5 @@
+require("./style.scss");
+
 function lectureItemController() {
     let self = this;
     let lectures = [{
@@ -18,10 +20,10 @@ function lectureItemController() {
       "courseId": "58cdebf63e7922e18b4bbf76",
       "name": "HTML5 + Forms",
       "theme": "HTML",
-      "lectureScheduledDate": "25/11/2016",
+      "lectureScheduledDate": "27/11/2016",
       "lectureScheduledTime": "18:00",
       "lectorId": "Volodymyr Sen",
-      "homeworkDeadline": "02/12/2016",
+      "homeworkDeadline": "04/12/2016",
       "assistants": "Vira Lytvyn, Stanislav Melnychek",
       "contentLecture": "https://goo.gl/4EN3hm",
       "contentPractice": "https://www.dropbox.com/s/y9ljazztqgd66io/Archive.zip?dl=0",
