+++ client/src/components/lectureItem/style.scss
