+++ client/src/components/personItem/personItem.html
@@ -1,20 +1,20 @@
-<div class="col m6 l4" ng-repeat="person in $ctrl.persons">
+<div class="col m6 l4">
 	<div class="card person">
 	    <div class="card-image waves-effect waves-block waves-light">
-	      <img class="activator" ng-src="{{person.avatar}}">
+	      <img class="activator" ng-src="{{$ctrl.person.avatar}}">
 	    </div>
 	    <div class="card-content">
 	      <span class="card-title activator grey-text text-darken-4">
-			{{person.firstName}} {{person.lastName}}
+			{{$ctrl.person.firstName}} {{$ctrl.person.lastName}}
 	      	<i class="material-icons right">more_vert</i>
 	      </span>
-	      <p><a href="#">{{person.email}}</a></p>
+	      <p class="email-container"><a href="#">{{$ctrl.person.email}}</a></p>
 	    </div>
 	    <div class="card-reveal">
 	      <span class="card-title grey-text text-darken-4">About:<i class="material-icons right">close</i></span>
 	      <p>
-	      	I am {{person.age}}.<br />
-	      	{{person.aboutMe}}
+	      	I am {{$ctrl.person.age}}.<br />
+	      	{{$ctrl.person.aboutMe}}
 	      </p>
 	    </div>
 	</div>
