+++ client/src/components/personItem/personItemController.js
@@ -1,0 +1,6 @@
+require("./style.scss");
+
+function UsersListController() {
+}
+
+module.exports = UsersListController;
