+++ client/src/components/personItem/style.scss
@@ -1,0 +1,4 @@
+.person .email-container{
+  text-align: center;
+  font-size: 14px;
+}
