--- client/src/components/personsList/style.scss
@@ -1,3 +1,0 @@
-.person {
-	min-height: 410px
-}
