+++ client/src/states/student/index.html
@@ -10,15 +10,15 @@
             </div>
         </li>
         <li>
-            <a ui-sref="home">
+            <a ui-sref="student.home">
                 <i class="material-icons">pageview</i>Main page</a>
         </li>
         <li>
-            <a ui-sref="lectures">
+            <a ui-sref="student.lectures">
                 <i class="material-icons">today</i>Lectures</a>
         </li>
         <li>
-            <a ui-sref="teachers">
+            <a ui-sref="student.teachers">
                 <i class="material-icons">toc</i>List teachers</a>
         </li>
         <li>
