+++ client/src/states/student/studentList/style.scss
@@ -1,0 +1,3 @@
+.person {
+  min-height: 410px
+}
