+++ client/src/states/student/studentList/template.html
@@ -1,29 +1,5 @@
 <div class="headerTop">
-  <h1>Students List</h1>
+  <h1>STUDENTS LIST</h1>
 </div>
 
-<div class="main">
-<div class="col s9">
-    <div class="row">
-        <div class="col s4" ng-repeat="user in $ctrl.users">
-            <div class="card">
-                <div class="card-image waves-effect waves-block waves-light">
-                    <img class="activator" ng-src="{{user.avatar}}" height="50%">
-                </div>
-                <div class="card-content">
-                    <span class="card-title activator grey-text text-darken-4">{{user.firstName}} {{user.lastName}}
-                        <i class="material-icons right">more_vert</i></span>
-                    <p><a ng-href="mailto:{{user.email}}">{{user.email}}</a></p>
-                </div>
-                <div class="card-reveal">
-                    <span class="card-title grey-text text-darken-4">About me:<i class="material-icons right">close</i></span>
-                    <p>
-                        I am {{user.age}}.<br />
-                        {{user.aboutMe}}
-                    </p>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-</div>
+<person-item  ng-repeat="person in $ctrl.persons" person="person"></person-item>
