+++ client/src/style.scss
@@ -7,9 +7,10 @@ body {
 }
 
 .headerTop {
-  margin: 24px 0;
+  margin: 25px 0;
 
   h1 {
+    font-weight: bold;
     color: #464547;
     font-size: 32px;
     text-align: center;
